[hir2mpl]Internal commit msg:     [hir2mpl] fix safe warning file line missing     [hir2mpl] add pointer array assertnonnull checking [mapleall]Internal commit msg:     [mapleall] pclint waring clean12

+++ src/mapleall/maple_be/include/cg/aarch64/aarch64_memlayout.h
@@ -151,7 +151,7 @@ class AArch64MemLayout : public MemLayout {
 
   SymbolAlloc *AssignLocationToSpillReg(regno_t vrNum) override;
 
-  uint32 StackFrameSize();
+  uint64 StackFrameSize();
 
   uint32 RealStackFrameSize();
